tests: make tests runnable without AbuseFilter
On CI, we would like to remove the recursive injection of dependencies.

CheckUser has integration tests with AbuseFilter which is otherwise not
needed by extensions depending on CheckUser. We have several of them:
- IPInfo
- LoginNotify
- Nuke
- OATHAuth
- WikimediaEvents
- WikimediaMaintenance

When running tests without AbuseFilter, BatchTemporaryAccountHandlerTest
would cause a PHP Fatal error as AbuseFilter FilterFromSpecsTestTrait
can not be found.

Héctor Arroyo must have encountered a similar issue and introduced a
proxy with Ia1df603e3ac0f6f213e0a6f0f78bf7fb31c56049. The last sentence
of its inline comment is:

> The anonymous class returned by this method allows for calling the
> private method from the trait, while allowing for not trying to load
> the trait at all if AbuseFilter is not loaded.

Which exactly matches my use case.  I have moved Héctor method from
CheckUserExpiredIdsLookupServiceTest to trait FilterFactoryProxyTrait
and use it in BatchTemporaryAccountHandlerTest for the same purpose.

The addDBDataForAbuseLog() is skipped since AbuseFilter is not loaded.
When it runs, it initializes $expiredAFLogId and $unavailableAFLogId
which are used in the data provider. Without AbuseFilter they are left
uninitialized and PHP complains about it, I am thus initializing them
to an arbitrary `-1`.

I have tested this using the following configurations:

Without AbuseFilter:

  wfLoadExtensions( [
  	'CheckUser',
  	'GlobalPreferences',
  ] );

With AbuseFilter:

  wfLoadExtensions( [
  	'CheckUser',
  	'GlobalPreferences',
  ] );

And for each configuration using following command:

  vendor/bin/phpunit extensions/CheckUser

This will make it possible to stop injecting AbuseFilter when a project
solely depends on CheckUser.

Bug: T389998
Change-Id: I9d226a67842bda6f7a0a43e9abc097bf3727c333