Load IPInfo message conditionally
On CI, we would like to remove the recursive injection of extensions
dependencies. When CheckUser is a dependency, we are required to add
IPInfo to fullfil the requirement that ipinfo-preference-use-agreement
message is defined.

This makes the message to be added solely when IPInfo is loaded.

Move tempAccountOnboarding registration from `extension.json` to the PHP
`ResourceLoaderRegisterModulesHook` hook. Add the ipinfo message solely
when IPInfo is loaded.

Bug: T389998
Change-Id: Iabcb934d472c9ed9c24a347ed7ea0c7580ff111e