Load IPInfo message conditionally
On CI, we would like to remove the recursive injection of extensions
dependencies. When CheckUser is a dependency, we are required to add
IPInfo to fullfil the requirement that ipinfo-preference-use-agreement
message is defined.

This makes the message to be added solely when IPInfo is loaded.

Move tempAccountOnboarding registration from `extension.json` to the PHP
`ResourceLoaderRegisterModulesHook` hook. Add the ipinfo message solely
when IPInfo is loaded.

I have resynced the extension.json since REL1_44..master had several
changes.

Bug: T389998
Change-Id: Iabcb934d472c9ed9c24a347ed7ea0c7580ff111e
(cherry picked from commit 635be4d88edd98f7c5027ebd06fa9665866df76a)